+++ src/integTest/projects/grails-project/src/test/groovy/issue75/MyControllerSpec.groovy
@@ -1,0 +1,25 @@
+package issue75
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
+ */
+@TestFor(MyController)
+class MyControllerSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        when:
+            def result = controller.index()
+
+        then:
+            result == 'hit main controller branch'
+    }
+}